+++ src/graphql/resolvers/mutations/delete-todo.js
