+++ src/graphql/resolvers/resolvers.js
@@ -1,9 +1,8 @@
-// import { withFilter, AuthenticationError } from 'apollo-server';
-import { createNewTodo } from './mutations/createNewTodo'
-import { getTodos } from './queries/getTodos'
-import { updateTodo } from './mutations/updateTodo'
-import { updateTodoStatus } from './mutations/updateTodoStatus'
-import { deleteTodo } from './mutations/deleteTodo'
+import { createNewTodo } from './mutations/create-new-todo'
+import { getTodos } from './queries/get-todos'
+import { updateTodo } from './mutations/update-todo'
+import { updateTodoStatus } from './mutations/update-todo-status'
+import { deleteTodo } from './mutations/delete-todo'
 import todoAdded from './subscriptions/todo-added'
 import todoDeleted from './subscriptions/todo-deleted'
 import todoUpdated from './subscriptions/todo-updated'
